requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
sqlmodel>=0.0.16
pymysql>=1.1.0
cryptography>=42.0.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

yfinance>=0.2.40
pandas>=2.2.0
numpy>=1.26.0
pandas-ta>=0.3.14b0

feedparser>=6.0.11
httpx>=0.27.0
tenacity>=8.2.3
python-dateutil>=2.9.0.post0

apscheduler>=3.10.4
passlib[bcrypt]>=1.7.4
pyjwt>=2.10.0
